Split files by backend type
